+++ sdk/src/core-plugins.ts
@@ -26,6 +26,7 @@ export { uiMaskPlugin, UIMaskPlugin } from './ui/render/mask';
 export { uiInteractionPlugin, UIInteractionPlugin } from './ui/input/interaction';
 export { uiLayoutPlugin, UILayoutPlugin } from './ui/layout/layout';
 export { uiRenderOrderPlugin, UIRenderOrderPlugin } from './ui/render/render-order';
+export { uiVisibilityPlugin, UIVisibilityPlugin } from './ui/core/visibility';
 export { textInputPlugin, TextInputPlugin } from './ui/text/text-input-plugin';
 export { dragPlugin, DragPlugin } from './ui/input/drag';
 export { focusPlugin, FocusPlugin } from './ui/input/focus';